It's been too long... Mainly recoded UIChatbar: Lost some features but a much better platform to work off of

+++ Character/OtherChar.h
@@ -28,29 +28,29 @@
 
 namespace ms
 {
-	// Other client's players.
+	// Other client players
 	class OtherChar : public Char
 	{
 	public:
-		OtherChar(int32_t charid, const CharLook& look, uint8_t level, int16_t job, const std::string& name, int8_t stance, Point<int16_t> position);
+		OtherChar(int32_t charid, const CharLook& look, uint16_t level, int16_t job, const std::string& name, int8_t stance, Point<int16_t> position);
 
-		// Update the character.
+		// Update the character
 		int8_t update(const Physics& physics) override;
-		// Add the movements which this character will go through next.
+		// Add the movements which this character will go through next
 		void send_movement(const std::vector<Movement>& movements);
 
-		// Update a skill level.
+		// Update a skill level
 		void update_skill(int32_t skillid, uint8_t skilllevel);
-		// Update the attack speed.
+		// Update the attack speed
 		void update_speed(uint8_t attackspeed);
-		// Update the character look.
+		// Update the character look
 		void update_look(const LookEntry& look);
 
-		// Return the character's attacking speed.
+		// Return the character's attacking speed
 		int8_t get_integer_attackspeed() const override;
-		// Return the character's level.
+		// Return the character's level
 		uint16_t get_level() const override;
-		// Return the character's level of a skill.
+		// Return the character's level of a skill
 		int32_t get_skilllevel(int32_t skillid) const override;
 
 	private:
